Correct For New DIV in Jekyll Syntax Markup
Jekyll 3.6.2 appears to introduce an extra div tag wrapping syntax highlight
blocks.  Locally testing with Jekyll 3.4.4 does not reproduce the issue
but updating to 3.6.2 demonstrates it.  It's also visible on the
production github pages site.

This new markup results in syntax blocks that can appear overly tall when
aligned with the base of the "On This Page" navigation section.

This commit introduces a CSS workaround that fixes the visual issue but
does not address the extra markup being put into the rendered HTML.