Add note to comparisons about data typing
Took me forever to realize that  #stage('Validation')['status'] == 'SUCCEEDED' wasn't working because the status is an enum internally, not a string (even though in the JSON it's represented as a string.  Solution was to add .toString() to it.  This seemed like such an inobvious solution/problem that it warranted a note in the docs.