fix(provider/kubernetes): remove deployment edit from RBAC instructions (#1049)
The GKE RBAC guide seems to have gotten a little mixed up here.

From reading through the guide I don't think that `my-test-account` is
intended to run Spinnaker itself. The user is required to already be
running a Spinnaker instance somewhere, per the prerequisites.

The intention is to help a user configure a new account, named
`my-test-account`, that has RBAC configured correctly so that Spinnaker
can deploy resources into it.