docs(ssl): update expected openssl inputs (#1626)
* docs(ssl): update expected openssl inputs

Commands provided will not prompt user for the password as it is already specified in the OpenSSL arguments. Updating documentation to make it much clearer to the user what option is being performed.

* docs(security): update expected openssl inputs
Commands provided will not prompt user for the password as it is already specified in the OpenSSL arguments. Updating documentation to make it much clearer to the user what option is being performed.