refactor(changelog): Revert changelog links (#1748)
* refactor(changelog): Update changelog links

To simplify the build process, rather than create an entirely
new gist for every release, we'll only create a gist for each
minor Spinnaker version (1.18.x, 1.19.x, etc.). That gist will
contain the release notes for each patch release within that
minor version as separate files (ex: 1.18.2.md, 1.19.1.md).

I've created these gists for the 1.17, 1.18 and 1.19 branches;
this points the docs site to these new gists. By passing the
file URL parameter we will only embed the specific file related
to the patch release in question, which means this change will
result in no difference from a user's perspective; it just
simplifies creating these gists.

* refactor(changelog): Revert changelog links

It turns out there's a problem with using a single gist
for all of the changelogs in a minor version; the maximum
size that will be rendered for a gist is across all of the
files in a gist, and the 1.17.7 release pushed the 1.17.x
gist across that size limit.  This means that trying to view
the 1.17.7 release notes now just yields a message that the
gist can't be rendered and a link to view the raw gist.

Reverting this as we'll need a different approach (maybe that
involves less copy-pasting of the content so the gist doesn't
grow as N^2 in the number of patch releases).

This reverts commit b8084fbba1b842e2c6d4b6720ba9aa0f3ab17fbc.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>