chore(build): Simplify changelog gists
The prior attempt to simplify changelog gists tried
putting all the changelogs for a particular minor release
as different files in the same gist.  This failed because
for some releases we exceed the size that can be rendered
for a gist (which causes failed rendering even if the
particular file we're viewing is less than the max size).

The root cause was that we were copy-pasting the prior patch
release's changlog into the new changelog and appending changes
at the top.  This mean that, for example for 1.17.x, with 7 patch
releases, we had the 1.17.0 release notes copied 7 times.

Rather than manually copy these release notes from gist to gist,
let's just update the website to display the gists for the current
patch release and all prior patch releases in order. This means
that when making a new patch release, the gist just needs to
contain the notes for that patch; the site will then incorporate
the prior notes by reference. I have a corresponding update to
the buildtool to update how we generate these changelog files,
so they automatically pull all patch releases, but this PR is
to backfill the currently-supported releases.  (I actually
used the buildtool locally to generate this change.)

With this change, we greatly reduce the total size of the gist
for each release, getting around the earlier problem.