feat(oauth2): Improve OAuth 2.0 Documentation (#1833)
* feat(oauth2): Improve OAuth 2.0 Documentation

Currently there is a lot of logic baked into Halyard commands to
automatically set fields in the OAuth 2.0 config. While in some
cases it might make set-up easier, it comes at the expense of
the ability to see and modify what is being configured.

Kleat will not replicate this behavior; it will not store a mapping
of the token URI for a number of common OAuth 2.0 providers; users
will need to set this themselves in their YAML config. In order
to make this easier, let's improve the documentation so that
it's easy for users to set up any OAuth 2.0 provider by just
pasting in the provided YAML and configuring as necessary.

I think this will actually make it easier for users to troubleshoot
as it will be more clear to them what fields they have set (rather
than having this done behind the scenes).

This commit:
(1) Adds full documentation on the OAuth 2.0 config; it is currently
copy-pasted from the source of truth protos, but once kleat is more
mature we can probably just link out to its documentation. (I don't
want a consumer of the documentation yet so we are more free to refactor
it as we'll likely think about once all the protos are written).
(2) Augments all the inline hal commands with documentation on how to
configure something by editing your YAML. (I actually think this will be
more clear to operators familiar with YAML than the hal commands.)
(3) Updates the provider-specific documentation to include a block of
YAML including what Halyard defaulted and kleat will not default.
(4) Updates references to all say OAuth 2.0 for consistency (fixing
the mix of OAuth2 Oauth2 OAuth 2 and OAuth 2.0).

Users using Halyard can use either the YAML-editing or CLI instructions;
users using kleat will need to use the YAML-editing instructions. (Though
given that kleat is not even alpha yet I have not mentioned it at all
here, we'll need to holistically update the docs when it's ready for users.)

* Apply suggestions from code review

Co-authored-by: Maggie Neterval <mneterval@google.com>
Co-authored-by: Dave Dorbin <ddorbin@google.com>

* style(oauth2): Address code review comments

Co-authored-by: Maggie Neterval <mneterval@google.com>
Co-authored-by: Dave Dorbin <ddorbin@google.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>