changed the title and a 1 line description (#1934)
* changed the title and a 1 line description

I changed the title to include the name of the AWS EKS tutorial
I added a 1 line description. The AWS EC2 provider is part of the AWS Cloud Provider and 
is used to manage AWS EC2 instances.
Hopefully this will help in my goal to re-accomodate this doc.

* Update setup/install/providers/aws/aws-ec2.md

thanks. I agree.

Co-authored-by: Dave Dorbin <ddorbin@google.com>

Co-authored-by: Dave Dorbin <ddorbin@google.com>