Update index.md
Suggestion to simplify a sentence to make it shorter.
Another suggestion I have is to add additional content to the sentence "It's a good idea though to communicate to the team what templates are available." but I don't exactly know the real reason why it is good.

Is it because it will save time for others having to recreate something that may already exist? Sharing why it's good may encourage people to communicate more.