feat(docs): slim down annotation-driven UI contract
Since we are committing to a contract in which the annotation-driven UI is supplied the full Kubernetes manifest to hydrate template values, let's encourage users to read values readily available there rather than relying on them as top-level keys for each infrastructure type. Specifically, let's remove apiVersion, displayName, kind, and namespace, as these are guaranteed to be available on the manifest, and Deck is currently just reading these from the manifest anyway to populate these fields. Alternately, we could have Clouddriver include these fields rather than having Deck read them directly off the manifest, but since we are trying to minimize Clouddriver's contract with Deck, let's do this instead.

Also removes the Instance type's `hasHealthStatus`, which does not appear to be getting set by Clouddriver anyway.