Update issues bulletpoint and fix a typo
I updated the link in what-to-hack.md to point to spinnaker/spinnaker repo and added a direct link to issues in that repo. I also removed an extra bullet from an item in the list.

These changes were needed because the original link pointed to the Spinnaker organization page on GitHub and you can't access the issues we were referring to from that page.