Updating deprecated SQL connection pool property (#2052)
After enabling DEBUG logging on orca, I discovered some of my SQL Connection pool properties were not being applied because sql.connectionPool has been deprecated in favor of sql.connectionPools, and some pool properties have changed names. The most misleading example in the current version is the maxLifetime property, which took me some days to realize was not having any effect, because it has been changed to maxLifetimeMs.

In order to more easily lead people to the latest reference, I changed the configuration example to be more in line with the ones I found for Front50 and CloudDriver, linking to the properties in kork-sql.