Update quote font style and add white space between case studies
I changed the featured quote text to be italicized,  vertically centered with the image, and consistent with the size of the case study text.

I added a gap between the rows of the case studies because there wasn't much space between one of the "learn" links and the study below it.